+++ src/main/java/com/academy/mudogroupware/file/application/command/GeneratePresignedUploadUrlCommand.java
@@ -1,7 +1,6 @@
 package com.academy.mudogroupware.file.application.command;
 
 public record GeneratePresignedUploadUrlCommand(
-        Long academyId,
         String fileName,
         String contentType
 ) {